requirements.txt
# backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
alembic==1.12.1
authlib==1.2.1
itsdangerous==2.1.2
stripe==7.9.0
starlette==0.27.0